[Issue-80] Add logic filter utxos for transfer
